cffif/requirements.txt
python-telegram-bot==20.6
Pillow==10.0.0